Abstract `LaserDevice`. Add `NannouLaserDevice`.
This commit abstracts the `LaserDevice` into its own abstract class
along with the LASER `Point` type and a single virtual method called
`addLineSequence` (previously `concatFramePoints`). See the method
documentation comment for details on requirements for implementation.

A new `NannouLaserDevice` class has been added that inherits from
`LaserDevice`. The `LaserDeviceFactory` and `LaserDeviceIdentifier`
classes have been renamed to `NannouLaserDeviceFactory` and
`NannouLaserDeviceIdentifier` respectively.

The `Renderer` types remain generic over all classes inheriting from
`LaserDevice`.

This commit also removes some magic variables in favour of documented
static const fields.
